WATCHPROP_CHANGESPROP constant; other changes
